+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -10,7 +10,7 @@ public class StatsServiceTest {
         StatsService service = new StatsService();
 
         long[] sales = {2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13};
-        long expected = 90;
+        int expected = 90;
 
         long actual = service.summSales(sales);
 
@@ -23,7 +23,7 @@ public class StatsServiceTest {
         StatsService service = new StatsService();
 
         long[] sales = {2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13};
-        long expected = 7;
+        int expected = 7;
 
         long actual = service.awgSales(sales);
 
@@ -35,9 +35,9 @@ public class StatsServiceTest {
         StatsService service= new StatsService();
 
         long[] sales = {2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13};
-        long expected = 12;
+        int expected = 12;
 
-        long actual = service.maxSales(sales);
+        int actual = service.maxSales(sales);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -47,9 +47,9 @@ public class StatsServiceTest {
         StatsService service = new StatsService();
 
         long[] sales = {2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13};
-        long expected = 1;
+        int expected = 1;
 
-        long actual = service.minSales(sales);
+        int actual = service.minSales(sales);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -59,9 +59,9 @@ public class StatsServiceTest {
         StatsService service = new StatsService();
 
         long[] sales = {2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13};
-        long expected = 5; // На самом деле месяцев 6. Дробные значения округляет в меньшую сторону, поэтому один месяц может выпасть
+        int expected = 5; // На самом деле месяцев 6. Дробные значения округляет в меньшую сторону, поэтому один месяц может выпасть
 
-        long actual = service.lessAwgSales(sales);
+        int actual = service.lessAwgSales(sales);
 
         Assertions.assertEquals(expected, actual);
     }
@@ -71,9 +71,9 @@ public class StatsServiceTest {
         StatsService service = new StatsService();
 
         long[] sales = {2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13};
-        long expected = 6;
+        int expected = 6;
 
-        long actual = service.aboveAwgSales(sales);
+        int actual = service.aboveAwgSales(sales);
 
         Assertions.assertEquals(expected, actual);
     }
